refactor: 将模型导入路径统一迁移到类型目录
将多个组件中的模型导入路径从相对路径改为统一的@/types路径，提高代码维护性
